deep_research/requirements.txt
openai
pydantic
python-dotenv
google-search-results
matplotlib
langchain-community
unstructured
